Changed scrolling control to the segment. This makes more sense since the segment should control the components inside
